Python/Create_DR/requirements.txt
pandas
inflect
python-docx
openpyxl
matplotlib
markitdown